+++ EventService/src/main/java/com/dwellora/dto/EventResponseDTO.java
@@ -2,13 +2,13 @@ package com.dwellora.dto;
 
 import java.time.LocalDateTime;
 
+/** Data transfer object containing event response details. */
 public record EventResponseDTO(
-        Integer eventId,
-        Integer apartmentId,
+        Long eventId,
+        Long apartmentId,
         String title,
         String description,
         LocalDateTime eventDate,
         Integer capacity,
         Integer currentRsvps,
-        boolean isFull
-) {}
+        boolean isFull) {}
