+++ MaintenanceService/src/main/java/com/dwellora/dto/UserDTO.java
@@ -1,21 +1,53 @@
 package com.dwellora.dto;
 
+/** Data transfer object representing user profile information. */
 public class UserDTO {
-    private Integer userId;
-    private Integer apartmentId;
+
+    private Long userId;
+    private Long apartmentId;
     private String fullName;
     private String flatNumber;
     private String role;
 
     public UserDTO() {}
-    public Integer getUserId() { return userId; }
-    public void setUserId(Integer userId) { this.userId = userId; }
-    public Integer getApartmentId() { return apartmentId; }
-    public void setApartmentId(Integer apartmentId) { this.apartmentId = apartmentId; }
-    public String getFullName() { return fullName; }
-    public void setFullName(String fullName) { this.fullName = fullName; }
-    public String getFlatNumber() { return flatNumber; }
-    public void setFlatNumber(String flatNumber) { this.flatNumber = flatNumber; }
-    public String getRole() { return role; }
-    public void setRole(String role) { this.role = role; }
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Long userId) {
+        this.userId = userId;
+    }
+
+    public Long getApartmentId() {
+        return apartmentId;
+    }
+
+    public void setApartmentId(Long apartmentId) {
+        this.apartmentId = apartmentId;
+    }
+
+    public String getFullName() {
+        return fullName;
+    }
+
+    public void setFullName(String fullName) {
+        this.fullName = fullName;
+    }
+
+    public String getFlatNumber() {
+        return flatNumber;
+    }
+
+    public void setFlatNumber(String flatNumber) {
+        this.flatNumber = flatNumber;
+    }
+
+    public String getRole() {
+        return role;
+    }
+
+    public void setRole(String role) {
+        this.role = role;
+    }
 }
