+++ NoticeService/src/main/java/com/dwellora/repository/NoticeRepository.java
@@ -1,15 +1,17 @@
 package com.dwellora.repository;
 
 import com.dwellora.entity.Notice;
+import java.time.LocalDateTime;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDateTime;
-import java.util.List;
-
+/** Repository interface for executing database operations on Notice entities. */
 @Repository
-public interface NoticeRepository extends JpaRepository<Notice, Integer> {
+public interface NoticeRepository extends JpaRepository<Notice, Long> {
 
-    List<Notice> findByApartmentIdAndExpiresAtAfterOrApartmentIdAndExpiresAtIsNullOrderByPublishedAtDesc(
-            Integer apartmentId1, LocalDateTime now, Integer apartmentId2);
+    /** Retrieves active notices that have not expired or have no expiration date. */
+    List<Notice>
+    findByApartmentIdAndExpiresAtAfterOrApartmentIdAndExpiresAtIsNullOrderByPublishedAtDesc(
+            Long apartmentId1, LocalDateTime now, Long apartmentId2);
 }
