+++ NotificationService/src/main/java/com/dwellora/event/BookingCreatedEvent.java
@@ -3,11 +3,14 @@ package com.dwellora.event;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
+/**
+ * Event published when an amenity booking is created.
+ */
 public class BookingCreatedEvent {
 
-    private Integer bookingId;
-    private Integer userId;
-    private Integer amenityId;
+    private Long bookingId;
+    private Long userId;
+    private Long amenityId;
     private String amenityName;
     private LocalDate bookingDate;
     private LocalTime startTime;
@@ -15,27 +18,27 @@ public class BookingCreatedEvent {
 
     public BookingCreatedEvent() {}
 
-    public Integer getBookingId() {
+    public Long getBookingId() {
         return bookingId;
     }
 
-    public void setBookingId(Integer bookingId) {
+    public void setBookingId(Long bookingId) {
         this.bookingId = bookingId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getAmenityId() {
+    public Long getAmenityId() {
         return amenityId;
     }
 
-    public void setAmenityId(Integer amenityId) {
+    public void setAmenityId(Long amenityId) {
         this.amenityId = amenityId;
     }
