+++ NotificationService/src/main/java/com/dwellora/service/NotificationService.java
@@ -3,9 +3,18 @@ package com.dwellora.service;
 import com.dwellora.dto.NotificationResponseDTO;
 import java.util.List;
 
+/**
+ * Service interface defining operations for retrieving and managing user notifications.
+ */
 public interface NotificationService {
 
-    List<NotificationResponseDTO> getUserNotifications(Integer userId);
+    /**
+     * Retrieves all notifications associated with a specific user ID.
+     */
+    List<NotificationResponseDTO> getUserNotifications(Long userId);
 
-    NotificationResponseDTO markAsRead(Integer notificationId);
+    /**
+     * Marks a specific notification as read.
+     */
+    NotificationResponseDTO markAsRead(Long notificationId);
 }
